third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/skip-waiting-installed.https-expected.txt

This is a testharness.js-based test.
FAIL Test skipWaiting when a installed worker is waitingassert_equals: skipWaiting promise should be resolved with undefined expected "PASS" but got "FAIL: Promise should be resolved before worker is activated"
Harness: the test ran to completion.
